Moved CampusPlateApi to Repository; Added dynamic endpoint resolution; Added Indicator for no listings or reservations; Moved Repository logic to companion object; Added Check for invalid credentials

+++ app/src/main/java/edu/cwru/caslab/campusplate/model/Listing.kt
@@ -9,7 +9,7 @@ data class Listing(
   val creationTime: Long,
   val expirationTime: Long,
   val quantity: Int,
-  val weightOunces: Int,
+  val weightOunces: Double,
   val quantityRemaining: String,
   var image: String?
 )
@@ -25,7 +25,7 @@ data class ListingCreationRequest(
   val title: String,
   val description: String,
   val quantity: Int,
-  val weightOunces: Int,
+  val weightOunces: Double,
   val creationDate: Long,
   val expirationDate: Long,
   val image: String?
